Finding the current location of this developer through geolocation by getting the ip address of this developer where cities are stored in GeoIP2 database and gets the city based on the ip address of the computer
Note: To get to the ASP.Net Core geolocation web application and get past the loopback address that ends in .1, e.g. 127.0.0.1,  install the ModHeader extension from the Chrome Web Store and input an IP address with the request header being named and entered as X-Forwarded for if using the application and downloading it from my Github account.